rename config packages
