+++ apps/docs/package.json
@@ -25,10 +25,10 @@
 	"devDependencies": {
 		"@eslint/compat": "^1.2.5",
 		"@eslint/js": "^9.18.0",
-		"@ldn-viz/eslint-config": "*",
+		"@ldn-viz/config-eslint": "*",
+		"@ldn-viz/config-typescript": "*",
+		"@ldn-viz/config-vitest": "*",
 		"@ldn-viz/themes": "*",
-		"@ldn-viz/typescript-config": "*",
-		"@ldn-viz/vitest-config": "*",
 		"@playwright/test": "^1.49.1",
 		"@storybook/addon-a11y": "^9.0.15",
 		"@storybook/addon-docs": "^9.0.15",
