+++ apps/docs/tsconfig.json
@@ -1,5 +1,5 @@
 {
-	"extends": ["@ldn-viz/typescript-config/svelte.json", "./.svelte-kit/tsconfig.json"],
+	"extends": ["@ldn-viz/config-typescript/svelte.json", "./.svelte-kit/tsconfig.json"],
 	"compilerOptions": {
 		"declaration": true,
 		"declarationMap": true,
