+++ packages/charts/vite.config.ts
@@ -1,7 +1,7 @@
 /// <reference types="vitest/config" />
 import { sveltekit } from '@sveltejs/kit/vite';
 import { defineConfig } from 'vite';
-import { sveltekitConfig, baseConfig } from '@ldn-viz/vitest-config';
+import { sveltekitConfig, baseConfig } from '@ldn-viz/config-vitest';
 
 export default defineConfig({
 	plugins: [sveltekit()],
