+++ packages/config-eslint/index.js
