+++ packages/config-typescipt/base.json
