+++ packages/config-typescipt/package.json
@@ -1,5 +1,5 @@
 {
-	"name": "@ldn-viz/typescript-config",
+	"name": "@ldn-viz/config-typescript",
 	"version": "0.0.0",
 	"private": true,
 	"license": "MIT",
