+++ packages/config-typescipt/svelte.json
