+++ packages/config-vitest/package.json
@@ -1,5 +1,5 @@
 {
-	"name": "@ldn-viz/vitest-config",
+	"name": "@ldn-viz/config-vitest",
 	"type": "module",
 	"exports": {
 		".": "./dist/index.js",
@@ -20,7 +20,7 @@
 		"vitest": "^3.2.4"
 	},
 	"devDependencies": {
-		"@ldn-viz/typescript-config": "*",
+		"@ldn-viz/config-typescript": "*",
 		"@vitest/coverage-v8": "^3.2.4",
 		"@vitest/ui": "^3.2.4",
 		"glob": "^11.0.1",
