+++ packages/config-vitest/src/configs/base-config.ts
