+++ packages/maps/package.json
@@ -40,9 +40,9 @@
 	"devDependencies": {
 		"@eslint/compat": "^1.2.5",
 		"@eslint/js": "^9.18.0",
-		"@ldn-viz/eslint-config": "*",
-		"@ldn-viz/typescript-config": "*",
-		"@ldn-viz/vitest-config": "*",
+		"@ldn-viz/config-eslint": "*",
+		"@ldn-viz/config-typescript": "*",
+		"@ldn-viz/config-vitest": "*",
 		"@playwright/test": "^1.49.1",
 		"@sveltejs/adapter-static": "^3.0.8",
 		"@sveltejs/kit": "^2.16.0",
