+++ packages/themes/eslint.config.js
@@ -1,3 +1,3 @@
-import { config } from '@ldn-viz/eslint-config/index.js';
+import { config } from '@ldn-viz/config-eslint/index.js';
 
 export default [...config];
